We have scraped Data from Family Dollar Website. Then we used the data to model our coupons.
As of now the program looking good. It fulfills the design process picture and has most to everything the base program would need. Im not quite sure what to do from here, but im guessing we can start scraping the real deal and try implementing it into the program.

We should also add some functionality to that bottom box

But that is all for now, good job